CLN: Move get_spark_testing_client to the spark backend
